--- src/components/InputBar/index.js
@@ -1,1 +1,0 @@
-export { default as InputBar } from './InputBar';
