+++ src/components/index.js
@@ -1,12 +1,10 @@
 export * from './BoardCard';
 export * from './Card';
 export * from './Carousel';
-export * from './Comment';
 export * from './Filter';
 export * from './Fieldset';
 export * from './Filter';
 export * from './Input';
-export * from './InputBar';
 export * from './ListHeader';
 export * from './MainPageListItem';
 export * from './PopUp';
