+++ src/feature/comment/component/index.js
@@ -1,6 +1,8 @@
-export { default as Comment } from './Comment';
+export { default as Comment } from './Comment/Comment';
 export { default as NestedComment } from './NestedComment/NestedComment';
 
 export { default as Comments } from './Comments/Comments';
 export { default as CommentsSuspense } from './Comments/CommentsSuspense';
 export { default as CommentsErrorFallback } from './Comments/CommentsErrorFallback';
+
+export { default as CommentInput } from './CommentInput/CommentInput';
