+++ src/feature/comment/hook/index.js
@@ -1,0 +1,1 @@
+export { default as useComment } from './useComment';
