+++ src/feature/comment/lib/index.js
@@ -1,0 +1,2 @@
+export * from './comment.js';
+export * from './filterComment.js';
