+++ src/hooks/index.js
@@ -1,4 +1,3 @@
-export { default as useComment } from './useComment';
 export { default as useLike } from './useLike';
 export { default as usePopUp } from './usePopUp.jsx';
 export { default as useScrap } from './useScrap';
