+++ src/hooks/useLike.jsx
@@ -1,12 +1,13 @@
 import { useParams } from 'react-router-dom';
 import { useQueryClient, useMutation } from '@tanstack/react-query';
 
+import { like as likeApi, unlike as unlikeApi } from '@/apis';
+
 import { useToast } from '@/shared/hook';
 import { flatPaginationCache, toPaginationCacheFormat } from '@/shared/lib';
 import { MUTATION_KEY, QUERY_KEY, LIKE_TYPE } from '@/shared/constant';
 
-import { like as likeApi, unlike as unlikeApi } from '@/apis';
-import { updateLikeIfTargetComment } from '@/utils';
+import { updateLikeIfTargetComment } from '@/feature/comment/lib';
 
 export default function useLike({ type, sourceId }) {
   const queryClient = useQueryClient();
