+++ src/utils/index.js
@@ -1,3 +1,1 @@
-export * from './comment.js';
 export * from './convertDS.js';
-export * from './filterComment.js';
